% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3ExcelObjects.R
\name{as.data.frame.excel_sheet}
\alias{as.data.frame.excel_sheet}
\title{Coerce excel_sheet to data.frame}
\usage{
\method{as.data.frame}{excel_sheet}(x, ...)
}
\arguments{
\item{x}{An excel_sheet object.}

\item{...}{Additional arguments.}
}
\description{
Coerce excel_sheet to data.frame
}
