% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3ExcelObjects.R
\name{get_data}
\alias{get_data}
\title{Extract Data from an excel_sheet}
\usage{
get_data(x)
}
\arguments{
\item{x}{An \code{excel_sheet} object.}
}
\value{
A data.frame or tibble.
}
\description{
Returns the imported data stored inside an \code{excel_sheet} object.
}
