% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3ExcelObjects.R
\name{get_filename}
\alias{get_filename}
\title{Get File Path from excel_file}
\usage{
get_filename(x)
}
\arguments{
\item{x}{An object of class \code{excel_file}.}
}
\value{
Character scalar containing the file path.
}
\description{
Extracts the physical file location.
}
