% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3ExcelObjects.R
\name{get_sheet}
\alias{get_sheet}
\title{Extract Sheet from excel_book}
\usage{
get_sheet(book, name)
}
\arguments{
\item{book}{An \code{excel_book} object.}

\item{name}{Character scalar or numeric position.}
}
\value{
An \code{excel_sheet} object or NULL.
}
\description{
Returns a worksheet from an imported workbook.
}
