% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3ExcelObjects.R
\name{names.excel_sheet}
\alias{names.excel_sheet}
\title{Names of data}
\usage{
\method{names}{excel_sheet}(x)
}
\arguments{
\item{x}{An excel_sheet object.}
}
\value{
Character vector
}
\description{
Names of data
}
