% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3ExcelObjects.R
\name{ncol.excel_sheet}
\alias{ncol.excel_sheet}
\title{Number of Columns in an excel_sheet}
\usage{
ncol.excel_sheet(x)
}
\arguments{
\item{x}{An excel_sheet object.}
}
\value{
Integer.
}
\description{
Number of Columns in an excel_sheet
}
