% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3ExcelObjects.R
\name{new_excel_book}
\alias{new_excel_book}
\title{Create a New excel_book Object}
\usage{
new_excel_book(sheets = list(), failed_sheets = list(), file = NULL)
}
\arguments{
\item{sheets}{Named list of \code{excel_sheet} objects.}

\item{failed_sheets}{Named list of failed worksheet objects.}

\item{file}{Optional \code{excel_file} object.}
}
\value{
An object of class \code{excel_book}.
}
\description{
Internal constructor for an \code{excel_book} S3 object.
}
\details{
An \code{excel_book} contains imported worksheets and failed worksheet
information.
}
\keyword{internal}
