% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3ExcelObjects.R
\name{new_excel_file}
\alias{new_excel_file}
\title{Create a New excel_file Object}
\usage{
new_excel_file(filename)
}
\arguments{
\item{filename}{Character scalar. Path to the Excel file.}
}
\value{
An object of class \code{excel_file}.
}
\description{
Internal constructor for an \code{excel_file} S3 object.
}
\details{
The object represents the physical Excel workbook stored on disk,
independently from its imported content.
}
\keyword{internal}
