% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3ExcelObjects.R
\name{new_excel_sheet}
\alias{new_excel_sheet}
\title{Create a New excel_sheet Object}
\usage{
new_excel_sheet(data, sheet, source)
}
\arguments{
\item{data}{A data.frame or tibble.}

\item{sheet}{Character scalar. Worksheet name.}

\item{source}{Character scalar. Excel file path.}
}
\value{
An object of class \code{excel_sheet}.
}
\description{
Internal constructor for an \code{excel_sheet} S3 object.
}
\details{
An \code{excel_sheet} represents a single worksheet imported from an
Excel workbook together with basic metadata describing its origin.
}
\keyword{internal}
