% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3ExcelObjects.R
\name{nrow.excel_sheet}
\alias{nrow.excel_sheet}
\title{Number of Rows in an excel_sheet}
\usage{
nrow.excel_sheet(x)
}
\arguments{
\item{x}{An excel_sheet object.}
}
\value{
Integer.
}
\description{
Number of Rows in an excel_sheet
}
