% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3ExcelObjects.R
\name{print.excel_book}
\alias{print.excel_book}
\title{Print excel_book Object}
\usage{
\method{print}{excel_book}(x, ...)
}
\arguments{
\item{x}{An object of class \code{excel_book}.}

\item{...}{Additional arguments.}
}
\description{
Displays workbook summary information.
}
