% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3ExcelObjects.R
\name{print.excel_file}
\alias{print.excel_file}
\title{Print excel_file Object}
\usage{
\method{print}{excel_file}(x, ...)
}
\arguments{
\item{x}{An object of class \code{excel_file}.}

\item{...}{Additional arguments.}
}
\description{
Displays a compact summary of an Excel file.
}
