% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadExcelHelpers.R
\name{select_sheets}
\alias{select_sheets}
\title{Select Workbook Sheets}
\usage{
select_sheets(all_sheets, sheet_names = NULL)
}
\arguments{
\item{all_sheets}{Character vector with available sheets.}

\item{sheet_names}{Optional requested sheets.}
}
\value{
Character vector with selected sheet names.
}
\description{
Selects worksheets from an available workbook structure.
}
\keyword{internal}
