% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3ExcelObjects.R
\name{sheet_names}
\alias{sheet_names}
\title{List Sheets in excel_book}
\usage{
sheet_names(x)
}
\arguments{
\item{x}{An \code{excel_book} object.}
}
\value{
Character vector with sheet names.
}
\description{
Returns available worksheet names.
}
