% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{xlversion-package}
\alias{xlversion}
\alias{xlversion-package}
\title{xlversion: Versioning and change tracking for Excel files (.xlsx)}
\description{
xlversion provides tools to create fingerprints, snapshots, and intelligent comparisons of Excel files at the sheet level. Ideal for version control, auditing, regulatory compliance and data pipelines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Verag/xlversion}
  \item Report bugs at \url{https://github.com/Verag/xlversion/issues}
}

}
\author{
\strong{Maintainer}: Vera Gonçalves \email{vera.maria.dg@gmail.com}

Authors:
\itemize{
  \item Vera Gonçalves \email{vera.maria.dg@gmail.com}
}

}
\keyword{internal}
